pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vid2cc-ai"
version = "0.1.0"
description = "AI-powered subtitle generator using Whisper and FFmpeg"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai-whisper",
    "ffmpeg-python",
    "torch",
    "tqdm",
]

[project.scripts]
vid2cc = "vid2cc_ai.cli:main"